Proper naming
An opcode (operation code) is a machine level instruction. In our case an opcode takes a byte (8 bits), the actual cpu instruction is located in the most significant nibble (4 bits) and an operand can be located in the least significant nibble. This means we can only have 16 different operations and the value of an operand can be from 0 to 15 (0-F).